Started refactoring the code
