+++ ap_credentials.json
@@ -1,0 +1,4 @@
+{
+    "ssid" : "",
+    "password" : ""
+}
