+++ main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-"""
+'''
 The MIT License (MIT)
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -24,7 +24,7 @@ Copyright (c) 2018 Mauro Riva (lemariva.com) for MicroPython on ESP32 and Amazon
 For a complete discussion, see http://www.makermusings.com
 More info about the MicroPython Version see https://lemariva.com
 
-"""
+'''
 import gc
 import machine
 import network
@@ -45,10 +45,6 @@ try:
     import usocket as socket
 except:
     import socket
-try:
-    import ustruct as struct
-except:
-    import struct
 
 # for ws2812b
 from wipyWS2812.ws2812 import WS2812
@@ -62,119 +58,13 @@ try:
 except:
     from ESP32MicroPython.timeutils import RTC
 
-# This XML is the minimum needed to define one of our virtual switches
-# to the Amazon Echo Dot / Amazon Echo (2nd generation)
+from xmls import SETUP_XML, EVENT_SERVICE_XML, GET_BINARY_STATE_SOAP
+from helpers import dbg, inet_aton, format_timetuple_and_zone
+from global_data import *
+from rest_api_handler import RestApiHandler
+from upnp_broadcast_responder import UpnpBroadcastResponder
 
-SETUP_XML = """<?xml version="1.0"?>
-<root>
-  <device>
-    <deviceType>urn:LeMaRiva:device:controllee:1</deviceType>
-    <friendlyName>%(device_name)s</friendlyName>
-    <manufacturer>Belkin International Inc.</manufacturer>
-    <modelName>Emulated Socket</modelName>
-    <modelNumber>3.1415</modelNumber>
-    <UDN>uuid:Socket-1_0-%(device_serial)s</UDN>
-    <serialNumber>221517K0101769</serialNumber>
-    <binaryState>0</binaryState>
-    <serviceList>
-      <service>
-          <serviceType>urn:Belkin:service:basicevent:1</serviceType>
-          <serviceId>urn:Belkin:serviceId:basicevent1</serviceId>
-          <controlURL>/upnp/control/basicevent1</controlURL>
-          <eventSubURL>/upnp/event/basicevent1</eventSubURL>
-          <SCPDURL>/eventservice.xml</SCPDURL>
-      </service>
-    </serviceList>
-  </device>
-</root>
-"""
-
-eventservice_xml = """<?scpd xmlns="urn:Belkin:service-1-0"?>
-<actionList>
-  <action>
-    <name>SetBinaryState</name>
-    <argumentList>
-      <argument>
-        <retval/>
-        <name>BinaryState</name>
-        <relatedStateVariable>BinaryState</relatedStateVariable>
-        <direction>in</direction>
-      </argument>
-    </argumentList>
-     <serviceStateTable>
-      <stateVariable sendEvents="yes">
-        <name>BinaryState</name>
-        <dataType>Boolean</dataType>
-        <defaultValue>0</defaultValue>
-      </stateVariable>
-      <stateVariable sendEvents="yes">
-        <name>level</name>
-        <dataType>string</dataType>
-        <defaultValue>0</defaultValue>
-      </stateVariable>
-    </serviceStateTable>
-  </action>
-</scpd>
-"""
-
-GetBinaryState_soap = """<s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/" s:encodingStyle="http://schemas.xmlsoap.org/soap/encoding/">
-  <s:Body>
-    <u:GetBinaryStateResponse xmlns:u="urn:Belkin:service:basicevent:1">
-      <BinaryState>%(state_realy)s</BinaryState>
-    </u:GetBinaryStateResponse>
-  </s:Body>
-</s:Envelope>
-"""
-
-DEBUG = True
-
-INADDR_ANY = 0
-global_epoch = 0  # time over ntp-server
-
-# W2812b
-ledNumber = 144  # number of leds
-chain = []
-clock = None
-
-
-def dbg(msg):
-    global DEBUG
-    if DEBUG:
-        print(msg)
-
-
-def inet_aton(addr):
-    ip_as_bytes = bytes(map(int, addr.split(".")))
-    return ip_as_bytes
-
-
-def format_timetuple_and_zone(timetuple, zone):
-    return "%s, %02d %s %04d %02d:%02d:%02d %s" % (
-        ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"][timetuple[6]],
-        timetuple[2],
-        [
-            "Jan",
-            "Feb",
-            "Mar",
-            "Apr",
-            "May",
-            "Jun",
-            "Jul",
-            "Aug",
-            "Sep",
-            "Oct",
-            "Nov",
-            "Dec",
-        ][timetuple[1] - 1],
-        timetuple[0],
-        timetuple[3],
-        timetuple[4],
-        timetuple[5],
-        zone,
-    )
-
-
-class poller:
+class Poller:
     # A simple utility class to wait for incoming data to be
     # read on a socket.
     def __init__(self):
@@ -222,25 +112,25 @@ class poller:
                 target.do_read(one_ready[0])
 
 
-class upnp_device:
-    """
+class Upnp_device:
+    '''
      Base class for a generic UPnP device. This is far from complete
      but it supports either specified or automatic IP address and port
      selection.
-    """
+    '''
 
     this_host_ip = None
 
     @staticmethod
     def local_ip_address():
-        if not upnp_device.this_host_ip:
+        if not Upnp_device.this_host_ip:
             try:
                 ap_if = network.WLAN()
-                upnp_device.this_host_ip = ap_if.ifconfig()[0]
+                Upnp_device.this_host_ip = ap_if.ifconfig()[0]
             except:
-                upnp_device.this_host_ip = "127.0.0.1"
-            dbg("got local address of %s" % upnp_device.this_host_ip)
-        return upnp_device.this_host_ip
+                Upnp_device.this_host_ip = "127.0.0.1"
+            dbg("got local address of %s" % Upnp_device.this_host_ip)
+        return Upnp_device.this_host_ip
 
     def __init__(
         self,
@@ -265,7 +155,7 @@ class upnp_device:
         if ip_address:
             self.ip_address = ip_address
         else:
-            self.ip_address = upnp_device.local_ip_address()
+            self.ip_address = Upnp_device.local_ip_address()
 
         self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         self.socket.bind((self.ip_address, self.port))
@@ -354,16 +244,16 @@ class upnp_device:
             dbg("Got problem to send response %s" % str(e))
 
 
-class fauxmo(upnp_device):
-    """
+class Fauxmo(Upnp_device):
+    '''
      This subclass does the bulk of the work to mimic a WeMo switch on the network.
-    """
+    '''
 
     @staticmethod
     def make_uuid(name):
         return "".join(
             ["%x" % sum([ord(c) for c in name])]
-            + ["%x" % ord(c) for c in "%sfauxmo!" % name]
+            + ["%x" % ord(c) for c in "%sFauxmo!" % name]
         )[:14]
 
     def __init__(self, name, listener, poller, ip_address, port, action_handler=None):
@@ -373,7 +263,7 @@ class fauxmo(upnp_device):
         self.relayState = 0
         persistent_uuid = "Socket-1_0-" + self.serial
         other_headers = ["X-User-Agent: redsonic"]
-        upnp_device.__init__(
+        Upnp_device.__init__(
             self,
             listener,
             poller,
@@ -389,7 +279,7 @@ class fauxmo(upnp_device):
         else:
             self.action_handler = self
         dbg(
-            "FauxMo device '%s' ready on %s:%s"
+            "Fauxmo device '%s' ready on %s:%s"
             % (self.name, self.ip_address, self.port)
         )
 
@@ -402,7 +292,7 @@ class fauxmo(upnp_device):
             and data.find(b"urn:Belkin:service:basicevent:1#GetBinaryState") != -1
         ):
             state = self.getState()
-            soap = GetBinaryState_soap % {"state_realy": state}
+            soap = GET_BINARY_STATE_SOAP % {"state_realy": state}
             date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
             message = (
                 "HTTP/1.1 200 OK\r\n"
@@ -430,7 +320,7 @@ class fauxmo(upnp_device):
                 "X-User-Agent: redsonic\r\n"
                 "CONNECTION: close\r\n"
                 "\r\n"
-                "%s" % (len(eventservice_xml), date_str, eventservice_xml)
+                "%s" % (len(EVENT_SERVICE_XML), date_str, EVENT_SERVICE_XML)
             )
             socket.send(message)
         elif data.find(b"GET /setup.xml HTTP/1.1") == 0:
@@ -470,7 +360,7 @@ class fauxmo(upnp_device):
 
             if success:
                 state = self.getState()
-                soap = GetBinaryState_soap % {"state_realy": state}
+                soap = GET_BINARY_STATE_SOAP % {"state_realy": state}
                 date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
                 message = (
                     "HTTP/1.1 200 OK\r\n"
@@ -498,145 +388,6 @@ class fauxmo(upnp_device):
         return self.relayState
 
 
-class upnp_broadcast_responder:
-    """
-     Since we have a single process managing several virtual UPnP devices,
-     we only need a single listener for UPnP broadcasts. When a matching
-     search is received, it causes each device instance to respond.
-     Note that this is currently hard-coded to recognize only the search
-     from the Amazon Echo for WeMo devices. In particular, it does not
-     support the more common root device general search. The Echo
-     doesn't search for root devices.
-    """
-
-    TIMEOUT = 0
-    inprogress = False
-
-    def __init__(self):
-        self.devices = []
-
-    def init_socket(self):
-        ok = True
-        self.ip = "239.255.255.250"
-        self.port = 1900
-        try:
-            # This is needed to join a multicast group
-            self.mreq = struct.pack("4sl", inet_aton(self.ip), INADDR_ANY)
-            # Set up server socket
-            self.ssock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            self.ssock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-            try:
-                self.ssock.bind(("", self.port))
-            except Exception as e:
-                dbg("WARNING: Failed to bind %s:%d: %s", (self.ip, self.port, e))
-                ok = False
-            try:
-                dbg(
-                    "IP: "
-                    + str(socket.IPPROTO_IP)
-                    + " IP_ADD_MEMBERSHIP: "
-                    + str(socket.IP_ADD_MEMBERSHIP)
-                    + " mreq: "
-                    + str(self.mreq)
-                )
-                self.ssock.setsockopt(
-                    socket.IPPROTO_IP, socket.IP_ADD_MEMBERSHIP, self.mreq
-                )
-            except Exception as e:
-                dbg("WARNING: Failed to join multicast group!: " + str(e))
-                ok = False
-
-        except Exception as e:
-            dbg("Failed to initialize UPnP sockets!")
-            return False
-        if ok:
-            dbg("Listening for UPnP broadcasts")
-
-    def fileno(self):
-        return self.ssock.fileno()
-
-    def sockets(self):
-        return self.ssock
-
-    def do_read(self, fileno):
-        data, sender = self.recvfrom(1024)
-        if data:
-            # Issue https://github.com/kakopappa/arduino-esp8266-alexa-multiple-wemo-switch/issues/22
-            if data.find(b"M-SEARCH") == 0 and self.inprogress is False:
-                if (
-                    data.find(b"upnp:rootdevice") != -1
-                    or data.find(b"ssdp:all") != -1
-                    or data.find(b"urn:Belkin:device:**") != -1
-                ):
-                    for device in self.devices:
-                        time.sleep(0.5)
-                        device.respond_to_search(
-                            sender, "urn:Belkin:device:**"
-                        )  # (sender, 'upnp:rootdevice')?
-                        self.inprogress = True
-                else:
-                    pass
-            else:
-                pass
-
-    # Receive network data
-    def recvfrom(self, size):
-        if self.TIMEOUT:
-            self.ssock.setblocking(0)
-            ready = select.select([self.ssock], [], [], self.TIMEOUT)[0]
-        else:
-            self.ssock.setblocking(1)
-            ready = True
-
-        try:
-            if ready:
-                return self.ssock.recvfrom(size)
-            else:
-                return False, False
-        except Exception as e:
-            dbg(e)
-            return False, False
-
-    def add_device(self, device):
-        self.devices.append(device)
-        dbg("UPnP broadcast listener: new device registered")
-
-
-class rest_api_handler(object):
-    """
-     This is an example handler class. The fauxmo class expects handlers to be
-     instances of objects that have on() and off() methods that return True
-     on success and False otherwise.
-
-     This example class takes a color and brightness.
-     It returns always True and ignores any return data.
-    """
-
-    def __init__(self, on_color, on_brightness):
-        global ledNumber
-        global ws2812_chain
-        self.on_color = on_color
-        self.on_brightnessr = on_brightness
-        ws2812_chain.set_brightness(on_brightness)
-
-    def on(self):
-        global ledNumber
-        global ws2812_chain
-        # global_epoch = timeutils.epoch() # updating time using ntp
-        data = [self.on_color for i in range(ledNumber)]
-        ws2812_chain.show(data)
-        dbg("response on")
-        return True
-
-    def off(self):
-        global ledNumber
-        global ws2812_chain
-        # global_epoch = timeutils.epoch() # updating time using ntp
-        data = [(0, 0, 0) for i in range(ledNumber)]
-        ws2812_chain.show(data)
-        dbg("response off")
-        return True
-
 
 class InvalidPortException(Exception):
     # Exception definitions that are used in the package
@@ -650,13 +401,13 @@ def thread_echo(args):
     # Set up our singleton for polling the sockets for data ready
 
     ws2812_chain = WS2812(ledNumber=ledNumber, brightness=100)
-    p = poller()
+    p = Poller()
 
-    """
+    '''
      Each entry is a list with the following elements:
 
      # name of the virtual switch
-     # handler object with 'on' and 'off' methods (e.g. rest_api_handler((rrr, ggg, bbb), lux)})
+     # handler object with 'on' and 'off' methods (e.g. RestApiHandler((rrr, ggg, bbb), lux)})
      # port #
 
      NOTE: As of 2015-08-17, the Echo appears to have a hard-coded limit of
@@ -667,44 +418,44 @@ def thread_echo(args):
      list will be used.
      NOTE: Micropython has a limitation in the number of opened sockets (8).
      Then, the maximal device number is limited to 3.
-    """
+    '''
     devices = [
         {
             "description": "white led",
             "port": 12340,
-            "handler": rest_api_handler((255, 255, 255), 50),
+            "handler": RestApiHandler((255, 255, 255), 50),
         },
         {
             "description": "red led",
             "port": 12341,
-            "handler": rest_api_handler((255, 0, 0), 50),
+            "handler": RestApiHandler((255, 0, 0), 50),
         },
         {
             "description": "blue led",
             "port": 12342,
-            "handler": rest_api_handler((30, 144, 255), 90),
+            "handler": RestApiHandler((30, 144, 255), 90),
         },
         {
             "description": "green led",
             "port": 12343,
-            "handler": rest_api_handler((0, 255, 0), 90),
+            "handler": RestApiHandler((0, 255, 0), 90),
         },
         {
             "description": "orange led",
             "port": 12344,
-            "handler": rest_api_handler((255, 165, 0), 90),
+            "handler": RestApiHandler((255, 165, 0), 90),
         },
     ]
 
     # Set up our singleton listener for UPnP broadcasts
-    u = upnp_broadcast_responder()
+    u = UpnpBroadcastResponder()
     u.init_socket()
     # Add the UPnP broadcast listener to the poller so we can respond
     # when a broadcast is received.
     p.add(u)
 
-    # Create our FauxMo virtual switch devices
-    # Initialize FauxMo devices
+    # Create our Fauxmo virtual switch devices
+    # Initialize Fauxmo devices
     for device in devices:
         # if `port` doesn't exist, populate it
         # if it isnt an int, flip out with a descriptive exception
@@ -716,7 +467,7 @@ def thread_echo(args):
                     type(device["port"]), device["port"]
                 )
             )
-        fauxmo(
+        Fauxmo(
             device["description"],
             u,
             p,
